--- src/main/java/duke/Launcher.java
@@ -1,13 +1,0 @@
-package duke;
-
-import duke.gui.Main;
-import javafx.application.Application;
-
-/**
- * Launcher.
- */
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
