--- src/main/java/duke/command/DeleteCommand.java
@@ -1,39 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.task.Task;
-import duke.task.TaskList;
-import duke.ui.Ui;
-
-/**
- * Represents the Command Delete that deletes a task from the list.
- *
- * @author Joseph Oliver Lim
- */
-public class DeleteCommand extends Command {
-    private int index;
-
-    /**
-     * Constructs a DeleteCommand with the index of the task to be deleted.
-     *
-     * @param index The index of the task to be deleted.
-     */
-    public DeleteCommand(int index) {
-        this.index = index;
-    }
-
-    /**
-     * Executes the DeleteCommand.
-     *
-     * @param tasks The TaskList where the command is to be executed.
-     * @param ui The Ui that functions as user interface.
-     * @param storage The Storage that functions to store data.
-     * @return A String to be shown to the user after the command is executed.
-     */
-    @Override
-    public String execute(TaskList tasks, Ui ui, Storage storage) {
-        Task task = tasks.getTask(this.index);
-        tasks.removeTask(this.index);
-        return ui.deleteTask(task, tasks.getCountTasks());
-    }
-}
