--- src/main/java/duke/command/UnmarkCommand.java
@@ -1,39 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.task.Task;
-import duke.task.TaskList;
-import duke.ui.Ui;
-
-/**
- * Represents the Command Unmark that marks a task as not done yet.
- *
- * @author Joseph Oliver Lim
- */
-public class UnmarkCommand extends Command {
-    private int index;
-
-    /**
-     * Constructs an UnmarkCommand with the index of the task to be marked as not done yet.
-     *
-     * @param index The index of the task to be marked as not done yet.
-     */
-    public UnmarkCommand(int index) {
-        this.index = index;
-    }
-
-    /**
-     * Executes the UnmarkCommand.
-     *
-     * @param tasks The TaskList where the command is to be executed.
-     * @param ui The Ui that functions as user interface.
-     * @param storage The Storage that functions to store data.
-     * @return A String to be shown to the user after the command is executed.
-     */
-    @Override
-    public String execute(TaskList tasks, Ui ui, Storage storage) {
-        Task task = tasks.getTask(this.index);
-        task.unmarkAsDone();
-        return ui.unmarkAsDone(task);
-    }
-}
