+++ src/main/java/jeo/Launcher.java
@@ -1,0 +1,13 @@
+package jeo;
+
+import javafx.application.Application;
+import jeo.gui.Main;
+
+/**
+ * Launcher.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
