+++ src/main/java/jeo/gui/Main.java
@@ -1,0 +1,33 @@
+package jeo.gui;
+
+import java.io.IOException;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.scene.layout.AnchorPane;
+import javafx.stage.Stage;
+import jeo.Jeo;
+
+/**
+ * A GUI for Je-O chatbot using FXML.
+ */
+public class Main extends Application {
+
+    private Jeo jeo = new Jeo();
+
+    @Override
+    public void start(Stage stage) {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
+            AnchorPane ap = fxmlLoader.load();
+            Scene scene = new Scene(ap);
+            stage.setScene(scene);
+            fxmlLoader.<MainWindow>getController().setJeo(jeo);
+            stage.setTitle("Je-O Chatbot");
+            stage.show();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+}
