+++ src/main/java/jeo/gui/MainWindow.java
@@ -1,0 +1,62 @@
+package jeo.gui;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.VBox;
+import jeo.Jeo;
+import jeo.ui.Ui;
+
+/**
+ * Controller for MainWindow. Provides the layout for the other controls.
+ */
+public class MainWindow extends AnchorPane {
+    @FXML
+    private ScrollPane scrollPane;
+    @FXML
+    private VBox dialogContainer;
+    @FXML
+    private TextField userInput;
+    @FXML
+    private Button sendButton;
+
+    private Jeo jeo;
+
+    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
+    private Image jeoImage = new Image(this.getClass().getResourceAsStream("/images/DaJeo.png"));
+
+    /**
+     * Initializes the GUI.
+     */
+    @FXML
+    public void initialize() {
+        scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
+        Ui ui = new Ui();
+        dialogContainer.getChildren().addAll(
+                DialogBox.getJeoDialog(ui.hello(), jeoImage)
+        );
+    }
+
+    public void setJeo(Jeo j) {
+        jeo = j;
+    }
+
+    /**
+     * Creates two dialog boxes, one echoing user input and the other containing Jeo's reply and then appends them to
+     * the dialog container. Clears the user input after processing.
+     */
+    @FXML
+    private void handleUserInput() {
+        String input = userInput.getText();
+        String response = jeo.getResponse(input);
+        dialogContainer.getChildren().addAll(
+                DialogBox.getUserDialog(input, userImage),
+                DialogBox.getJeoDialog(response, jeoImage)
+        );
+        userInput.clear();
+    }
+}
+
