+++ src/main/resources/view/MainWindow.fxml
@@ -9,7 +9,7 @@
 <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity"
             minWidth="-Infinity" prefHeight="600.0" prefWidth="400.0"
             xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="duke.gui.MainWindow">
+            fx:controller="jeo.gui.MainWindow">
     <children>
         <TextField fx:id="userInput" layoutY="558.0" onAction="#handleUserInput"
                    prefHeight="41.0" prefWidth="324.0" AnchorPane.bottomAnchor="1.0" />
